Moved RemoteQueryCommand class to catalog.py
RemoteQueryCommand is a generic class which can be used to query
on remote hosts, Hence placing it in catalog.py where other functions
related to database query resides.
